refactor(tests): extract shared delivery helpers into _delivery_helpers.ts (closes #3075)
Move SAMPLE_ANALYSIS, seedDeliveryWithDecisions, seedProducts, PRODUCT_A_ID,
PRODUCT_B_ID, and makeInvoiceItem into tests/convex/_delivery_helpers.ts so
all three delivery test suites can share them without duplication.
